Latest version

+++ src/RandomIntegersGenerator.java
@@ -4,7 +4,7 @@ import java.util.Random;
  * Created by Łukasz on 2016-12-08.
  */
 public class RandomIntegersGenerator {
-    public static int[] generateIntsArray(int amount, int min, int max){
-        return new Random().ints(amount, min, max+1).toArray();
+    public static int[] generateIntsArray(long amount, long min, long max){
+        return new Random().ints(amount, (int)min, (int)max+1).toArray();
     }
 }
